Se integró la interfaz de ingreso al resto del paquete
