+++ SistemaDeVentas/src/App.jsx
@@ -3,20 +3,25 @@ import Index from 'pages';
 import EditAccount from 'pages/editAccount';
 import RegProd from 'pages/registro_producto';
 import {BrowserRouter as Router, Switch, Route} from "react-router-dom";
+import IngresoForm from 'pages/IngresoForm';
 
 function App() {
   return (
         <div className="App">
             <Router>
                 <Switch>
+                    
                     <Route path="/editaccount">
                         <EditAccount />
                     </Route>
                     <Route path="/registro_producto">
                         <RegProd/>
                     </Route>
+                    <Route path="/gestion-datos">
+                        <Index/>
+                    </Route>
                     <Route path="/">
-                        <Index />
+                    <IngresoForm />
                     </Route>
                 </Switch>
             </Router>
